+++ frontend/Idle/src/components/common/logos/MainLogoBlack.jsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { ReactComponent as MainLogoImg } from "../../assets/images/hyundai.svg";
+import { ReactComponent as MainLogoImg } from "../../../assets/images/hyundai.svg";
 import { useNavigate } from "react-router-dom";
 
 function MainLogoBlack() {
