+++ frontend/Idle/src/components/common/modals/WarningModal.jsx
@@ -1,8 +1,8 @@
 import styled from "styled-components";
 import { useState, useContext } from "react";
-import carContext from "../../utils/context";
-import BlueButton from "./BlueButton";
-import WhiteButton from "./WhiteButton";
+import carContext from "../../../utils/context";
+import BlueButton from "../buttons/BlueButton";
+import WhiteButton from "../buttons/WhiteButton";
 /**
  *
  * @param {string} title 질문내용 (문자열)
