+++ frontend/Idle/src/components/common/toolTips/FindTrimTooltip.jsx
