+++ frontend/Idle/src/components/trimBoxContainer/TrimBoxContainer.jsx
@@ -1,5 +1,5 @@
 import { styled } from "styled-components";
-import NormalTrimBox from "../common/NormalTrimBox";
+import NormalTrimBox from "../common/boxs/NormalTrimBox";
 
 function TrimBoxContainer(data) {
   return (
