+++ frontend/Idle/src/pages/trimPage.jsx
@@ -2,7 +2,7 @@ import { useContext, useEffect, useState } from "react";
 import { getTrimData } from "../utils/api";
 import TrimBoxContainer from "../components/trimBoxContainer/TrimBoxContainer";
 import { styled } from "styled-components";
-import BlueButton from "../components/common/BlueButton";
+import BlueButton from "../components/common/buttons/BlueButton";
 import { useNavigate } from "react-router-dom";
 import { TRIM_ROUTE } from "../utils/routes";
 import { carContext } from "../utils/context";
